[TACHYON-1140] Rename UnderStoreFileInStream to UnderStoreBlockInStream
