+++ clients/unshaded/src/main/java/tachyon/client/file/FileInStream.java
@@ -32,6 +32,7 @@ import tachyon.client.TachyonStorageType;
 import tachyon.client.block.BlockInStream;
 import tachyon.client.block.BufferedBlockOutStream;
 import tachyon.client.block.LocalBlockInStream;
+import tachyon.client.block.UnderStoreBlockInStream;
 import tachyon.client.file.options.InStreamOptions;
 import tachyon.exception.ExceptionMessage;
 import tachyon.exception.PreconditionMessage;
@@ -331,7 +332,7 @@ public final class FileInStream extends InputStream implements BoundedStream, Se
       }
       long blockStart = BlockId.getSequenceNumber(blockId) * mBlockSize;
       mCurrentBlockInStream =
-          new UnderStoreFileInStream(blockStart, mBlockSize, mFileInfo.getUfsPath());
+          new UnderStoreBlockInStream(blockStart, mBlockSize, mFileInfo.getUfsPath());
       mShouldCacheCurrentBlock = mTachyonStorageType.isStore();
     }
   }
